+++ backend/tests/test_epic_api.py
@@ -6,48 +6,13 @@ from sqlmodel import SQLModel, Session, create_engine
 from sqlmodel.pool import StaticPool
 from ..api.epic import get_session
 
-db_name = "test_db.sqlite"
-test_con = f"sqlite:///{db_name}"
-test_engine = create_engine(
-    test_con, connect_args={"check_same_thread": False}, echo=True
-)
 
-# Creates the DB and Deletes it after ALL tests are run thanks to scope=module
-@pytest.fixture(name="create_db", scope="module")
-def create_db():
-    # setup
-    SQLModel.metadata.create_all(test_engine)
-    yield
-    # teardown
-    os.remove(db_name)
-
-
-@pytest.fixture(name="session")
-def session_fixture(create_db):
-    create_db
-
-    with Session(test_engine) as session:
-        yield session
-
-
-@pytest.fixture(name="client")
-def client_fixture(session: Session):
-    def get_session_override():
-        return session
-
-    app.dependency_overrides[get_session] = get_session_override
-    client = TestClient(app)
-    yield client
-    app.dependency_overrides.clear()
-
-
+@pytest.mark.order(1)
 def test_post_epic(client):
     response = client.post(
         "/api/epics/",
         json={"name": "[dyvenia]branding", "work_area": "DYV", "client_id": 1},
     )
-    app.dependency_overrides.clear()
-
     data = response.json()
     assert response.status_code == 200
     assert data == {
@@ -91,6 +56,7 @@ def test_get_epic_list(client):
     ]
 
 
+@pytest.mark.order(-1)
 def test_delete_epics(client):
     response = client.delete("/api/epics/?epic_name=[dyvenia]branding")
     assert response.status_code == 200
